Create a new StanModel interface.
