+++ src/main/scala/com/cibo/scalastan/transform/Optimize.scala
@@ -10,10 +10,10 @@
 
 package com.cibo.scalastan.transform
 
-import com.cibo.scalastan.ScalaStan
+import com.cibo.scalastan.StanContext
 import com.cibo.scalastan.ast.StanProgram
 
-case class Optimize()(implicit val ss: ScalaStan) extends StanTransform[Unit] {
+case class Optimize()(implicit val context: StanContext) extends StanTransform[Unit] {
   def initialState: Unit = ()
 
   private val steps: Seq[StanTransform[_]] = Seq(
